Use tokens for session, fix reCaptcha
Switch from using cookies to using tokens for session logins.  This will require you to login again.  Also add a browser based worked around for reCaptchas.

Fix reCaptcha always being detected (because it's now mandatory on registration and the registration form is present on the login page).  This is an educated guess on my part as to where the recaptcha would appear if it was present for logging in, so if you have problems logging in, try logging in on the website and if the recaptcha pops up, please save a copy of the page and contact me.