Various Fixes / Additions
Don't treat 503 as a permanent error (GOG uses this as a temporarily overloaded error)

Add -covers and -backgrounds to download cover images and background images (respectively) for each game.

Change automatic cleanup of !downloading directory as part of download command to only delete stuff not in the manifest at all (instead of not in the current queue) to allow interrupting a download,  downloading something else and then resuming a partial download.

Add clear_partial_downloads command to remove everything from !downloading , in case something gets stuck/forgotten as a result of the previous.

Fixed failed to find token error message to check for both KeyError and AttributeErrror